Static LAG consistency fix (OVS vswitchd)
Fix issue with LAG configuration & unconfiguration.  In OVS, when a Port is
created with only one interface, it exists in h/w as simply a single
physical port.  When a second interface is added, the Port becomes a LAG in
h/w.  However, when the second interface is removed, the Port remains a LAG
with a single port in it.  Fixed this inconsistent behavior such that when
a Port (with LACP mode off) only has one configured interface, the LAG is
destroyed in h/w.

Change-Id: Ia48775b3cf826ece639976e5c863c480ef3c598d
Signed-off-by: Mike Chuang <mike.chuang@hp.com>