Support for on-demand fetching of non-monitored data
The IDL only support reading information from data that is part of the
replica. For processes that rarely access the data base, or part of it,
this could result in a waste of resources.

This commit adds support for reading data that is not being monitored
from the OVSDB server. The data can be retrived at row, column, table
level.

Change-Id: Ibfa403b8cdd00ee2ab54211df6f5b86ebba86cdd
Signed-off-by: Sebastian Arguello <sebastian.arguello@hpe.com>