new: dev: Add Partial Map Updates functionality
Add functionality to partially update a map column, which allows to update
key-value pairs inside the map column without updating the whole map column.

For each map column in the schema, two functions are generated:

ovsrec_<table_name>_update_<column_name>_setkey()
ovsrec_<table_name>_update_<column_name>_delkey()

The '_setkey()' functions can be used to insert a new key-value (if the key
doesn't exists in the map) or to change a value (if the key already exists in
the map). The '_delkey()' functions can be used to delete a key-value from the
map, using just the key.

Change-Id: Iaff288b293f53eb2b609958178995e16b36ce861
Signed-off-by: Edward Aymerich <edward.aymerich@hpe.com>