Asic-plugin.h is added into plugins directory
This commit is part of SwitchD Refactor
This change introduces the register of the asic plugin interfaces.
An asic plugin must provide to be compatible with SwitchD Asic
plugin infrastructure. When an external plugin attemps to register
itself as an Asic plugin, the code will validate that interface
provided meets the requirements for MAJOR and MINOR versions.

Change-Id: I5260a3be018a15ba878e015db2fdf4e290fc2252
Signed-off-by: Carolina Vargas Basilio <cbasilio@hpe.com>